apply mixin to all VueApplications
