+++ src/module/actor/sheets/sitesheet-v2.ts
@@ -49,21 +49,4 @@ export class IronswornSiteSheetV2 extends VueActorSheet {
     this.actor.update({ data: { denizens } }, { render: true })
     return true
   }
-
-  _getHeaderButtons() {
-    return [
-      {
-        class: 'ironsworn-toggle-edit-mode',
-        label: 'Edit',
-        icon: 'fas fa-edit',
-        onclick: (e) => this._toggleEditMode(e),
-      },
-      ...super._getHeaderButtons(),
-    ]
-  }
-
-  _toggleEditMode(_e: JQuery.ClickEvent) {
-    const currentValue = this.actor.getFlag('foundry-ironsworn', 'edit-mode')
-    this.actor.setFlag('foundry-ironsworn', 'edit-mode', !currentValue)
-  }
 }
