+++ src/module/applications/vueSfSettingTruthsDialog.ts
@@ -1,9 +1,10 @@
-import { VueApplication } from '../vue/vueapp'
 import { Starforged, starforged } from 'dataforged'
 import sfTruthsVue from '../vue/sf-truths.vue'
 import { VueSheetRenderHelperOptions } from '../vue/vue-render-helper'
+import { VueAppMixin } from '../vue/vueapp.js'
+import { App } from 'vue'
 
-export class SFSettingTruthsDialogVue extends VueApplication {
+export class SFSettingTruthsDialogVue extends VueAppMixin(Application) {
   static get defaultOptions(): ApplicationOptions {
     return mergeObject(super.defaultOptions, {
       title: game.i18n.localize('IRONSWORN.SFSettingTruthsTitle'),
@@ -17,7 +18,6 @@ export class SFSettingTruthsDialogVue extends VueApplication {
 
   get renderHelperOptions(): Partial<VueSheetRenderHelperOptions> {
     return {
-      ...super.renderHelperOptions,
       components: { 'sf-truths': sfTruthsVue },
       vueData: async () => ({
         // Avoid rollupjs's over-aggressive tree shaking
