+++ src/module/item/bondset/bondsetsheet-v2.ts
@@ -12,21 +12,4 @@ export class BondsetSheetV2 extends VueItemSheet {
       components: { 'bondset-sheet': bondsetSheetVue },
     }
   }
-
-  _getHeaderButtons() {
-    return [
-      {
-        class: 'ironsworn-toggle-edit-mode',
-        label: 'Edit',
-        icon: 'fas fa-edit',
-        onclick: (e) => this._toggleEditMode(e),
-      },
-      ...super._getHeaderButtons(),
-    ]
-  }
-
-  _toggleEditMode(_e: JQuery.ClickEvent) {
-    const currentValue = this.item.getFlag('foundry-ironsworn', 'edit-mode')
-    this.item.setFlag('foundry-ironsworn', 'edit-mode', !currentValue)
-  }
 }
