+++ src/module/item/move/sfmovesheet.ts
@@ -15,21 +15,4 @@ export class SFMoveSheet extends VueItemSheet {
       components: { 'sfmove-sheet': sfmoveSheetVue },
     }
   }
-
-  _getHeaderButtons() {
-    return [
-      {
-        class: 'ironsworn-toggle-edit-mode',
-        label: 'Edit',
-        icon: 'fas fa-edit',
-        onclick: (e) => this._toggleEditMode(e),
-      },
-      ...super._getHeaderButtons(),
-    ]
-  }
-
-  _toggleEditMode(_e: JQuery.ClickEvent) {
-    const currentValue = this.item.getFlag('foundry-ironsworn', 'edit-mode')
-    this.item.setFlag('foundry-ironsworn', 'edit-mode', !currentValue)
-  }
 }
