+++ src/module/item/progress/progresssheet-v2.ts
@@ -15,21 +15,4 @@ export class ProgressSheetV2 extends VueItemSheet {
       components: { 'progress-sheet': progressSheetVue },
     }
   }
-
-  _getHeaderButtons() {
-    return [
-      {
-        class: 'ironsworn-toggle-edit-mode',
-        label: 'Edit',
-        icon: 'fas fa-edit',
-        onclick: (e) => this._toggleEditMode(e),
-      },
-      ...super._getHeaderButtons(),
-    ]
-  }
-
-  _toggleEditMode(_e: JQuery.ClickEvent) {
-    const currentValue = this.item.getFlag('foundry-ironsworn', 'edit-mode')
-    this.item.setFlag('foundry-ironsworn', 'edit-mode', !currentValue)
-  }
 }
