+++ src/module/rolls/challenge-resolution-dialog.ts
@@ -1,8 +1,8 @@
 import { VueSheetRenderHelperOptions } from '../vue/vue-render-helper'
-import { VueApplication } from '../vue/vueapp'
 import VueDialog from '../vue/challenge-resolution-dialog.vue'
+import { VueAppMixin } from '../vue/vueapp.js'
 
-export class ChallengeResolutionDialog extends VueApplication {
+export class ChallengeResolutionDialog extends VueAppMixin(Application) {
   private constructor(
     protected messageId: string,
     options?: Partial<ApplicationOptions>
@@ -29,7 +29,7 @@ export class ChallengeResolutionDialog extends VueApplication {
     return this.openDialogs[messageId]
   }
 
-  close(options?: {}): Promise<void> {
+  close(options?: Application.CloseOptions): Promise<void> {
     delete ChallengeResolutionDialog.openDialogs[this.messageId]
     return super.close(options)
   }
