+++ src/module/vue/vueitemsheet.ts
@@ -1,12 +1,12 @@
 import { App } from 'vue'
 import { $ItemKey } from './provisions'
-import { VueMixin } from './vue-mixin.js'
+import { VueAppMixin } from './vueapp.js'
 import {
   VueSheetRenderHelper,
   VueSheetRenderHelperOptions,
 } from './vue-render-helper'
 
-export abstract class VueItemSheet extends VueMixin(ItemSheet) {
+export abstract class VueItemSheet extends VueAppMixin(ItemSheet) {
   renderHelper: VueSheetRenderHelper | undefined
 
   static get defaultOptions() {
@@ -40,8 +40,19 @@ export abstract class VueItemSheet extends VueMixin(ItemSheet) {
     return this
   }
 
-  close(options?: FormApplication.CloseOptions | undefined): Promise<void> {
-    this.renderHelper?.close()
-    return super.close(options)
+  _getHeaderButtons() {
+    return [
+      {
+        class: 'ironsworn-toggle-edit-mode',
+        label: 'Edit',
+        icon: 'fas fa-edit',
+        onclick: (e) => this._toggleEditMode(e),
+      },
+      ...super._getHeaderButtons(),
+    ]
+  }
+  _toggleEditMode(_e: JQuery.ClickEvent) {
+    const currentValue = this.item.getFlag('foundry-ironsworn', 'edit-mode')
+    this.item.setFlag('foundry-ironsworn', 'edit-mode', !currentValue)
   }
 }
